+++ tests/testthat/test_subcorpus.R
@@ -72,6 +72,42 @@ test_that(
   }
 )
 
+test_that(
+  "programming against subset()",
+  {
+    # corpus as point of departure
+    gparl <- corpus("GERMAPARLMINI")
+    persons <- c("Angela Dorothea Merkel", "Volker Kauder", "Ronald Pofalla")
+    subcorpora <- lapply(persons, function(who) subset(gparl, speaker == !!who))
+    names(subcorpora) <- persons
+    expect_identical(
+      unname(sapply(subcorpora, s_attributes, "speaker")),
+      persons
+    )
+    for (person in persons){
+      p <- partition("GERMAPARLMINI", speaker = person)
+      expect_identical(size(p), size(subcorpora[[person]]))
+    }
+    
+    g <- corpus("GERMAPARLMINI") %>% subset(quote(speaker == "Volker Kauder"))
+    expect_identical(size(g), size(subcorpora[["Volker Kauder"]]))
+    
+    # subcorpus as point of departure
+    gparl_sub <- corpus("GERMAPARLMINI") %>% subset(date == "2009-11-10")
+    persons <- c("Angela Dorothea Merkel", "Volker Kauder", "Ronald Pofalla")
+    subcorpora <- lapply(persons, function(who) subset(gparl_sub, speaker == !!who))
+    names(subcorpora) <- persons
+    for (person in persons){
+      p <- partition(as(gparl_sub, "partition"), speaker = person, verbose = FALSE)
+      expect_identical(size(p), size(subcorpora[[person]]))
+    }
+    
+    g <- corpus("GERMAPARLMINI") %>%
+      subset(quote(speaker == "Volker Kauder" & date == "2009-11-10"))
+    expect_identical(size(g), size(subcorpora[["Volker Kauder"]]))
+  }
+)
+
 
 test_that(
   "warnings if s-attributes are not valid",